Switched all external refs to rts2scriptelement.h to rts2script.h, so the template NextCommand function in rts2script.h works. The compiler was very grateful that I stuck (int) in front of a few more of P's enums....
git-svn-id: http://svn.code.sf.net/p/rts-2/code/branches/rts-2/REL_0_5_0_exp@3004 65ab8f4e-f147-0410-b3a9-f14133ecfe55

+++ src/plan/rts2scriptguiding.cpp
@@ -178,7 +178,7 @@ Rts2ScriptElementGuiding::nextCommand (Rts2DevClientCamera * camera,
   // should not happen!
   logStream (MESSAGE_DEBUG)
     << "Rts2ScriptElementGuiding::nextCommand invalid state: "
-    << processingState << sendLog;
+    << (int) processingState << sendLog;
   return NEXT_COMMAND_NEXT;
 }
 
@@ -188,13 +188,13 @@ Rts2ScriptElementGuiding::processImage (Rts2Image * image)
   int ret;
   Rts2ConnFocus *processor;
   logStream (MESSAGE_DEBUG) <<
-    "Rts2ScriptElementGuiding::processImage state: " << processingState <<
-    sendLog;
+    "Rts2ScriptElementGuiding::processImage state: " << (int) processingState
+    << sendLog;
   if (processingState != WAITING_IMAGE)
     {
       logStream (MESSAGE_ERROR)
 	<< "Rts2ScriptElementGuiding::processImage invalid processingState: "
-	<< processingState << sendLog;
+	<< (int) processingState << sendLog;
       processingState = FAILED;
       return -1;
     }
